feat: add R data processing and visualization pipeline
Introduce a new R-based data processing pipeline using the 'targets'
package. This includes functions for importing and plotting geographic
data, and a '_targets.R' script to define the workflow. Add Excel files
for raw station data and a '.gitignore' for R-related files to maintain
a clean repository. This setup facilitates efficient data management and
visualization, enhancing reproducibility and scalability of analyses.