docs: add README, LICENSE, and CODE_OF_CONDUCT files
Introduce a README.md to provide an overview of the project,
including features, data sources, usage instructions, and
acknowledgments. Add a LICENSE.md file to specify the MIT License
for the project. Include a CODE_OF_CONDUCT.md to establish community
standards and ensure a welcoming environment for contributors. Add
.Rbuildignore to exclude LICENSE.md from the build process.